breaking out time formatting functions into separate module, using to format seconds into time strings across app
